+++ spec/feature/signing_in_spec.rb
@@ -4,7 +4,6 @@ feature 'sign in' do
                  password: '12345678',
                  password_confirmation: '12345678')
   end
-
   scenario 'with correct credentials' do
     sign_in(email: maker.email,   password: maker.password)
     expect(page).to have_content "Welcome, #{maker.email}"
