+++ spec/helpers/session.rb
@@ -2,7 +2,7 @@ module SessionHelpers
   def sign_up(email: 'olivia@example.com',
           password: '12345678',
           password_confirmation: '12345678')
-    visit'/makers/new'
+    visit '/makers/new'
     fill_in :email, with: email
     fill_in :password, with: password
     fill_in :password_confirmation, with: password_confirmation
